+++ packages/plugin-collapsible-list/web/src/getToolbarButtons.tsx
@@ -2,6 +2,7 @@ import { PLUGIN_TOOLBAR_BUTTON_ID } from 'wix-rich-content-editor-common';
 import CollapsibleListSettings from './modals/SettingsModals';
 import { collapsibleModals } from './consts';
 import type { ToolbarButton } from 'ricos-types';
+import { TIPTAP_COLLAPSIBLE_LIST_TYPE } from 'ricos-content';
 
 export const getToolbarButtons = (config, services): ToolbarButton[] => {
   return [
@@ -20,12 +21,33 @@ export const getToolbarButtons = (config, services): ToolbarButton[] => {
           layout: isMobile ? 'fullscreen' : 'drawer',
         });
       },
+      attributes: {
+        selectedNode: selectedNodeResolver,
+      },
     },
     {
       id: PLUGIN_TOOLBAR_BUTTON_ID.SEPARATOR,
     },
     {
       id: PLUGIN_TOOLBAR_BUTTON_ID.DELETE,
+      attributes: {
+        selectedNode: selectedNodeResolver,
+      },
     },
   ];
+};
+
+const selectedNodeResolver = {
+  id: 'selectedNode',
+  resolve: (_, __, editor) => {
+    const { selection } = editor?.state || {};
+    const node =
+      selection?.from === selection?.to &&
+      selection?.$anchor?.path?.find(node => node?.type?.name === TIPTAP_COLLAPSIBLE_LIST_TYPE);
+    if (node) {
+      return node;
+    } else {
+      return undefined;
+    }
+  },
 };