+++ packages/plugin-commons/web/src/index.ts
@@ -36,3 +36,5 @@ export { default as ColorPicker } from './Components/ColorPicker/ColorPicker';
 export { getSelectionStyles } from 'wix-rich-content-editor-common';
 
 export { Uploader } from './Utils/Uploader';
+
+export { selectedNodeResolver } from './Utils/contentResolvers';