+++ packages/ricos-plugins/web/src/toolbarButtonsConfig.ts
@@ -10,7 +10,6 @@ import type { IToolbarItemConfigTiptap } from 'ricos-types';
 import { PLUGIN_TOOLBAR_BUTTON_ID } from 'wix-rich-content-editor-common';
 import { createLink } from 'ricos-content/libs/nodeUtils';
 import { convertRelObjectToString, convertRelStringToObject } from 'wix-rich-content-common';
-import type { ComponentType } from 'react';
 
 type PluginButtonId = typeof PLUGIN_TOOLBAR_BUTTON_ID[keyof typeof PLUGIN_TOOLBAR_BUTTON_ID];